agentic_codebase_genius/FE/requirements.txt
streamlit>=1.38
requests
python-dotenv